+++ src/components/App.tsx
@@ -5,15 +5,15 @@ import { AttachHistogramCallback, Histogram } from "./Histogram"
 import { Logquacious } from "../services/Logquacious"
 import { Query } from "../services/Query"
 import { ChangeFilterCallback, FilterDropdown, FilterItem } from "./FilterDropdown"
-import { ChangeSettingCallback, MenuSetting } from "./Menu/MenuSetting"
+import { ChangeSettingCallback, MenuSetting } from "./menu/MenuSetting"
 import { Direction, Theme } from "../services/Prefs"
 import { Welcome } from "./Welcome"
 import { Error } from "./Error"
 import { Title } from "./Title"
 import { SearchBar, SearchBarCallback } from "./SearchBar"
-import { ChangeRangeCallback, Picker } from "./picker/picker"
-import { MenuTitle } from "./Menu/MenuTitle"
-import { MenuDropdown } from "./Menu/MenuDropDown"
+import { ChangeRangeCallback, Picker } from "./picker/Picker"
+import { MenuTitle } from "./menu/MenuTitle"
+import { MenuDropdown } from "./menu/MenuDropDown"
 
 export type AttachResultsCallback = (el: HTMLElement) => void
 export type DisplayCallback = (d: Display, errorMessage?: string) => void
